+++ ios/RNTwilioVoice/RNTwilioVoice.m
@@ -16,6 +16,7 @@
 @property (nonatomic, strong) PKPushRegistry *voipRegistry;
 @property (nonatomic, strong) TVOCallInvite *callInvite;
 @property (nonatomic, strong) TVOCall *call;
+@property (nonatomic, strong) void(^callKitCompletionCallback)(BOOL);
 @property (nonatomic, strong) CXProvider *callKitProvider;
 @property (nonatomic, strong) CXCallController *callKitCallController;
 @end
@@ -42,7 +43,7 @@ RCT_EXPORT_MODULE()
 
 - (NSArray<NSString *> *)supportedEvents
 {
-  return @[@"connectionDidConnect", @"connectionDidDisconnect", @"callRejected", @"deviceReady"];
+  return @[@"connectionDidConnect", @"connectionDidDisconnect", @"callRejected", @"deviceReady", @"deviceNotReady"];
 }
 
 @synthesize bridge = _bridge;
@@ -82,12 +83,16 @@ RCT_EXPORT_METHOD(configureCallKit: (NSDictionary *)params) {
     NSLog(@"CallKit Initialized");
 
     self.callKitCallController = [[CXCallController alloc] init];
-    [self sendEventWithName:@"deviceReady" body:nil];
   }
 }
 
 RCT_EXPORT_METHOD(connect: (NSDictionary *)params) {
   NSLog(@"Calling phone number %@", [params valueForKey:@"To"]);
+
+//  [[TwilioVoice sharedInstance] setLogLevel:TVOLogLevelVerbose];
+
+  UIDevice* device = [UIDevice currentDevice];
+  device.proximityMonitoringEnabled = YES;
 
   if (self.call && self.call.state == TVOCallStateConnected) {
     [self.call disconnect];
@@ -213,9 +218,14 @@ RCT_REMAP_METHOD(getActiveCall,
                                                completion:^(NSError *error) {
                                                  if (error) {
                                                    NSLog(@"An error occurred while registering: %@", [error localizedDescription]);
+                                                   NSMutableDictionary *params = [[NSMutableDictionary alloc] init];
+                                                   [params setObject:[error localizedDescription] forKey:@"err"];
+
+                                                   [self sendEventWithName:@"deviceNotReady" body:params];
                                                  }
                                                  else {
                                                    NSLog(@"Successfully registered for VoIP push notifications.");
+                                                   [self sendEventWithName:@"deviceReady" body:nil];
                                                  }
                                                }];
   }
@@ -301,6 +311,9 @@ RCT_REMAP_METHOD(getActiveCall,
 #pragma mark - TVOCallDelegate
 - (void)callDidConnect:(TVOCall *)call {
   self.call = call;
+  self.callKitCompletionCallback(YES);
+  self.callKitCompletionCallback = nil;
+
   NSMutableDictionary *callParams = [[NSMutableDictionary alloc] init];
   [callParams setObject:call.sid forKey:@"call_sid"];
   if (call.state == TVOCallStateConnecting) {
@@ -333,10 +346,15 @@ RCT_REMAP_METHOD(getActiveCall,
     [params setObject:StateDisconnected forKey:@"call_state"];
   }
   [self sendEventWithName:@"connectionDidDisconnect" body:params];
-  if (self.call.state == TVOCallStateConnected) {
-    [self performEndCallActionWithUUID:call.uuid];
-  }
+
+  // if the app initiate and terminates the answered call
+  // EndCallAction inside performEndCallActionWithUUID will result into an error in the log
+  // because the CallKit reference is already ended.
+  // ref https://github.com/hoxfon/react-native-twilio-programmable-voice/pull/30
+  [self performEndCallActionWithUUID:call.uuid];
+
   self.call = nil;
+  self.callKitCompletionCallback = nil;
 }
 
 - (void)call:(TVOCall *)call didFailWithError:(NSError *)error {
@@ -360,9 +378,10 @@ RCT_REMAP_METHOD(getActiveCall,
   }
   [self sendEventWithName:@"connectionDidDisconnect" body:params];
 
-  if (self.call.state == TVOCallStateConnected) {
-    [self performEndCallActionWithUUID:call.uuid];
-  }
+  [self performEndCallActionWithUUID:call.uuid];
+
+  self.callKitCompletionCallback(NO);
+  self.callKitCompletionCallback = nil;
   self.call = nil;
 }
 
@@ -415,17 +434,19 @@ RCT_REMAP_METHOD(getActiveCall,
   NSLog(@"provider:performStartCallAction");
 
   [[TwilioVoice sharedInstance] configureAudioSession];
-  self.call = [[TwilioVoice sharedInstance] call:[self fetchAccessToken]
-                                          params:_callParams
-                                        delegate:self];
 
-  if (!self.call) {
-    [action fail];
-  } else {
-    self.call.uuid = action.callUUID;
+  [self.callKitProvider reportOutgoingCallWithUUID:action.callUUID startedConnectingAtDate:[NSDate date]];
 
-    [action fulfillWithDateStarted:[NSDate date]];
-  }
+  __weak typeof(self) weakSelf = self;
+  [self performVoiceCallWithUUID:action.callUUID client:nil completion:^(BOOL success) {
+    __strong typeof(self) strongSelf = weakSelf;
+    if (success) {
+      [strongSelf.callKitProvider reportOutgoingCallWithUUID:action.callUUID connectedAtDate:[NSDate date]];
+      [action fulfill];
+    } else {
+      [action fail];
+    }
+  }];
 }
 
 - (void)provider:(CXProvider *)provider performAnswerCallAction:(CXAnswerCallAction *)action {
@@ -436,12 +457,13 @@ RCT_REMAP_METHOD(getActiveCall,
   //      `provider:performAnswerCallAction:` per the WWDC examples.
   // [[TwilioVoice sharedInstance] configureAudioSession];
 
-  self.call = [self.callInvite acceptWithDelegate:self];
-  if (self.call) {
-    self.call.uuid = [action callUUID];
-  }
-
-  self.callInvite = nil;
+  [self performAnswerVoiceCallWithUUID:action.callUUID completion:^(BOOL success) {
+    if (success) {
+      [action fulfill];
+    } else {
+      [action fail];
+    }
+  }];
 
   [action fulfill];
 }
@@ -531,6 +553,38 @@ RCT_REMAP_METHOD(getActiveCall,
       NSLog(@"EndCallAction transaction request successful");
     }
   }];
+}
+
+- (void)performVoiceCallWithUUID:(NSUUID *)uuid
+                          client:(NSString *)client
+                      completion:(void(^)(BOOL success))completionHandler {
+
+    self.call = [[TwilioVoice sharedInstance] call:[self fetchAccessToken]
+                                            params:_callParams
+                                          delegate:self];
+
+    if (!self.call) {
+        completionHandler(NO);
+    } else {
+        self.call.uuid = uuid;
+    }
+
+    self.callKitCompletionCallback = completionHandler;
+}
+
+- (void)performAnswerVoiceCallWithUUID:(NSUUID *)uuid
+                            completion:(void(^)(BOOL success))completionHandler {
+
+    self.call = [self.callInvite acceptWithDelegate:self];
+    if (!self.call) {
+        completionHandler(NO);
+    } else {
+        self.call.uuid = uuid;
+    }
+
+    self.callInvite = nil;
+
+    self.callKitCompletionCallback = completionHandler;
 }
 
 @end